Move db creation scripts to dbutils
